chore: separate Shaderity class from index.ts
